Proposer rewrite.
A proposer can now handle instances that were preempted by another proposer. This commit also adds some changes to the acceptors, which now reply with "nacks", i.e. failed to promise or accept.
This commit also adds a couple of new test cases for the proposer. More tests are needed for instances that were preempted and that already have a value assigned.